Added limit switch query params
When the limit switch is called, the slave performs a status check on
the cell_id + 2^control_size. This activates the last control line
(highest power), which toggles to the limit switch in each cell as
opposed to it's sensor. There may be an issue handling analog vs.
digital in.