Implemented new column hookup, and echo.
Changed the logic surrounding adding a new address:
The master will periodically poll the default address with an ECHO
message, if it gets a response it will inform the front. If this
message is sent, it should be interpreted as there is a new column
ready for deployment, and logic to handle this should reside in the
front end.

Additionally, the slave now exclusively stores its own address in
memory. and always sets to DEFAULT if it is unable to read.