New Address Code is now std length
When directing a column to change its address, the "cell" value in the
command is used as the new address.

This is ultimately due to my profound disappointment with how
convoluted it is to append items to arrays without use of std:: etc.

Also it was a weird decision in the first place.