Impling Unlock and Queries
Unlock cycle now works. Master sends unlock cycle, and then enters a
loop where it continues querying the slave for limit switch status,
exitting the loop when the limit switch says the door is closed.

Function for door closed  - bool isDoorClosed(Message) - is factored out
to deal with active high or low limit switches (this depends on if we
use a pull up or down resistor).

Limit and analog reads now work, and pass correct values.